feat : í 회원가입 post, directory moved
