+++ backend/src/main/java/com/app/backend/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.app.backend.member.repository;
+package com.app.backend.domain.member.repository;
 
-import com.app.backend.member.entity.Member;
+import com.app.backend.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, String > {
